perf: pack Game storage into 6 merged slots (Stage B)
Behavior-preserving storage packing of DegenerusGameStorage + its modules:
- bingoFirsts: firstQuadrant+firstSymbol -> one uint64 mapping
- deityBoonPacked: deityBoonDay+deityBoonUsedMask -> one uint32 mapping
- levelDgnrsPacked: allocation+claimed -> one uint256 (128|128) + accessors
- DecClaimRound: 3 slots -> 1 (uint96 poolWei | uint128 totalBurn | uint32 rngWord); winner selection keeps the full VRF word, only the claim-time lootbox seed is narrowed (fed through keccak, never sliced)
- totalFlipReversals(uint64)+lastVrfProcessedTimestamp(uint48) share slot 5
- lootboxEvCapPacked: nested mapping -> one two-window mapping for the live {currentLevel, currentLevel+1} keys
RNG floor intact (reverseFlip rngLock gate + VRF freeze untouched; hash2 1-liner dropped as not byte-identical). Forge 849/0/110; 53 slot-hardcoded harnesses recalibrated to the POST layout (deleted slots 27/38/50/57). Game bytecode 19,403 (headroom 5,173).

+++ contracts/modules/DegenerusGameDecimatorModule.sol
@@ -270,9 +270,11 @@ contract DegenerusGameDecimatorModule is DegenerusGamePayoutUtils {
         emit DecimatorResolved(lvl, packedOffsets, poolWei, totalBurn);
 
         // Snapshot claim round for this level (persistent — no expiry)
-        round.poolWei = poolWei;
-        round.totalBurn = uint232(totalBurn);
-        round.rngWord = rngWord;
+        round.poolWei = uint96(poolWei);
+        round.totalBurn = uint128(totalBurn);
+        // Winners were already selected from the full VRF word above (decSeed) and packed
+        // into decBucketOffsetPacked; only the claim-time lootbox seed is stored, narrowed.
+        round.rngWord = uint32(rngWord);
 
         return 0; // All funds held for claims
     }
